pyproject.toml
[project]
name = "openshell-deep-agent"
version = "0.1.0"
description = "OpenShell agent with native Dapr durable-agent and LangGraph+Dapr runtimes"
requires-python = ">=3.12"
dependencies = [
    "dapr-agents==1.0.0",
    "dapr-ext-langgraph==1.17.3",
    "langchain>=1.2.12",
    "langgraph>=1.0.3",
    "openshell>=0.0.7",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = ["src"]
